Move reusable serialization code to a separate object
